Turn Threads synchronization
Now the Match, Round and Turn threads are in wait state while they are waiting for an event i.e. a player move.
So the Turn is waked up when a move is done, th Round when a turn is finished and the same is for the Round.
Every Thread has his specific attribute for calling wait() and notify() in order to be the only one to be notified.
Attributes :
- Match -> endRound
- Round -> hasEndedTurn
- Turn -> hasMade Move